Car type added

+++ Transport.h
@@ -1,8 +1,10 @@
 #include "Type.h";
 #include "Bus.h"
 #include "Truck.h"
+#include "Car.h"
+
 struct Transport
 {
 	type mKey;
-	short int tPower;
+	int tPower;
 };
